Fix Task shutdown and exit signal omission
Beforehand the exit status was read immediately when a channel close was
detected. This was wrong and resulted in sometimes the wrong exit status
being used. The exit status is only guaranteed to be available after the
channel is really closed on both sides. This was fixed with this commit.

http://b/201388753