Factor out timestamp and ordered_in_file_descriptor from PerfEvents
This is for performance.
Accessing the timestamp of these events is performance critical for
`PerfEventQueue` and the `Proc.Def.Events` thread.
Previously, this went through `std::visit` which is not as fast as expected.

We introduce a `struct PerfEvent` containing the `timestamp` (and
`ordered_in_file_descriptor` while we are here) so that it can be accessed
directly. The `std::variant` that was previously top-level is now *contained* in
`PerfEvent`, and the possible types are now called `...PerfEventData`.
An `Accept` method can now be added to wrap the calls to the remaining
occurrences of `std::visit`.
Unfortunately, the interface of `PerfEventVisitor` gets uglier because the
`std::variant` is no longer top level.

In order to keep the type of an event where it's known, a templated
`TypedPerfEvent` struct is introduced that contains only a specific
`...PerfEventData` instead of the variant. `using` declarations are added to
have back the previous `...PerfEvent` types.
These `...PerfEvent`s convert implicitly to the `PerfEvent` with the
`std::variant`.

Performance tested on Trata in two configurations:
1) Sampling 1000Hz, scheduling, uprobes on `item_coordinate_frame::update`;
2) Only uprobes on `item_coordinate_frame::update`.

In case 1, CPU utilization of `OrbitService` went down from 63-74 to 58-67,
and CPU utilization of `Proc.Def.Events` specifically from 41-48 to 36-42.
In case 2, CPU utilization of `OrbitService` went down from 29-36 to 28-33,
and CPU utilization of `Proc.Def.Events` specifically from 19-24 to 17-20.

Similar tests with uprobes on `execute_task` were done confirming the ~10%
relative improvement.

Test: Unit tests. `LinuxTracingIntegrationTests`. Various captures during
performance testing.