Improve ScopeTreeTimerDataTests
In this PR we are improving the tests in ScopeTreeTimerData, so each function
is properly tested.

In addition, we moved the logic of GetTimerMetadata method from the
provider to ScopeTreeTimerData, because it has more sense to have it
here. TimerMetadata appeared in the following doc
(http://go/stadia-orbit-timers-outside-ui) and aims to get
quickly all the data needed to draw in addition to the timer themselves.

This PR is a little piece inside http://b/202110929, aiming to have good
tests and no bugs when switching from the inner data to queries in the
provider (draft PR here: https://github.com/google/orbit/pull/2891).