Factor out timestamp and ordered_in_file_descriptor from PerfEvents
This is a change for performance that carries some cleanup with itself as a side
effect.

Accessing the timestamp of these events is performance critical for
`PerfEventQueue` and the `Proc.Def.Events` thread.
Previously, this went through `std::visit` which is not as fast as expected.

We introduce a `class PerfEvent` containing the `timestamp` (and
`ordered_in_file_descriptor` while we are here) so that it can be accessed
directly. The `std::variant` that was previously top-level is now *contained* in
`PerfEvent`, and the possible types are now called `...PerfEventData`.

In order to keep the type of an event where it's known (e.g.,
`Consume...PerfEvent` functions), a bit of non-virtual inheritance is
introduced. The templated `class TypedPerfEvent` exposes a `data()` accessor
that returns the expected type instead of the whole `std::variant`, without
overhead. Sub-types `...PerfEvent` represent the specific events and provide
dedicated constructors.
This was measured to have no effect on performance.

An `Accept` method is now added to wrap the calls to the remaining occurrences
of `std::visit`. This is necessary as the `std::visit` got way more complicated
and unfortunately contains a chain of (`constexpr`) `if`s and `static_cast`s,
but the additional complexity has to be somewhere and it's all internal to
`PerfEvent.h`/`.cpp`.
Compare https://github.com/google/orbit/pull/2910 or
https://github.com/google/orbit/pull/2911 for other possibilities that I believe
were less clean because they were exposing more of the complexity to the users
of the code.

Now that we have accessors, and `timestamp_` and `ordered_in_file_descriptor`
are private, we can remove `mutable` from some fields.

Performance tested on Trata in two configurations:
1) Sampling 1000Hz, scheduling, uprobes on `item_coordinate_frame::update`;
2) Only uprobes on `item_coordinate_frame::update`.

In case 1, CPU utilization of `OrbitService` went down from 63-74 to 58-67,
and CPU utilization of `Proc.Def.Events` specifically from 41-48 to 36-42.
In case 2, CPU utilization of `OrbitService` went down from 29-36 to 28-33,
and CPU utilization of `Proc.Def.Events` specifically from 19-24 to 17-20.

Similar tests with uprobes on `execute_task` were done confirming the ~10%
relative improvement.

Test: Unit tests. `LinuxTracingIntegrationTests`. Various captures during
performance testing.