Create TimerTrackDataIdManager
In order to be able to expand ThreadTrackDataProvider to also work with
other TimerTracks (GpuTrack, SchedulerTrack and FrameTrack) and therefore
rename it to TimerTrackDataProvider, we need an unique identifier
track_id (or pane_id in the future, see
go/stadia-orbit-timers-outside-ui) to know which timers should belong to
the same container.

Currently all this logic is inside TrackManager (UI), where we have different
maps for each type of Track and we use different functions to get the
Track themselves. Now, we are aiming to have a provider to get this data
and we also need that identifier to allow the Tracks (UI) to get their
data.

In this context, we are creating TimerTrackDataIdManager, who owns all
the existente track_data_id and has the main responsibility of identify
which timers should belong to the same data.

The main function is GetTrackIdFromTimerInfo, who only return the same
track_data_id for timers who should belong to the same container. We
are also providing GetTrackTypeId(track_type_identifier) for all timer
tracks to allow the possibility of creating a track without any timer.
This is mostly legacy and these functions might be private in the
future.

This PR is a little piece inside http://b/202110356, aiming to have an
only provider for all kind of TimerTracks.