Support ThreadTrack optimization in ScopeTreeTimerData
In this PR we are extending ScopeTreeTimerData to allow the optimization
made in https://github.com/google/orbit/pull/2780. In that PR, we
optimized ThreadTracks to no go through all the timers in a range if
after we were not going to draw them (we don't draw 2 different timer if
they occupt the same pixels).

So, here we add 2 methods to ScopeTreeTimerData class:

- GetTimersAtDepth (depth, min_timestamp, max_timestamp) : Get the timers
in one particular depth level.

- GetTimersAtDepthOptimized: (depth, resolution, min_timestamp,
max_timestamp): Get Timers in a range that after will be draw having
certain resolution.

This last method will be used by ThreadTrackDataProvider.

This PR is a puzzle piece of making ThreadTracks to use the provider
instead of their own data (https://github.com/google/orbit/pull/2891)
which aims to solve http://b/202110929.