Work with close/open intervals in GetTimers
We are assuming that intervals in GetTimers functions are close-open
[start-end) but it's more natural to ask for an open-open interval (all
timers inside the time-range), so now we change the internal methods to
adjust it.

Found when making new ScopeTreeTimerData tests but didn't want to mix
with that PR. Not a big issue because after will be filtered, so I
didn't change it in ThreadTracks, as we are changing this in a following
PR anyways.

The test was failing because the nextPixelBoundary was wrong, and
FindFirstScopeAtTimer was taking a previous pixel and therefore we were
getting an extra timer we shouldn't have got.