Compute capture section size correctly using also section list offset
The capture (main) section in our captures files extends from the
capture section offset to either the next section or to the section list
offset. So far, the section list offset was not considered when
computing the size of the capture section. This is now fixed.

The changed implementation does not change behavior for code that
does not read past the CaptureFinished event, however behavior for
reading past this event is now well specified: you get empty messages
until you hit the end of the capture section. Previously it was possible
to incorrectly parse data from the section list as proto messages.

Tested: Unit tests.