Integrate libunwindstack fork into the buildsystem
This is replacing the conan package of libunwindstack by the fork in
`third_party/`.

The fork will link against the new `libunwindstack-android-dependencies`
package.

The whole libunwindstack test suite is also integrated into our unit
test system.

So far only the really necessary changes to the codebase have been made.
We can consider making some more which can reduce the number of hacks
that needed to be done in the CMakeLists.txt file (like disabling
specific warning flags or the dependency of the tests to the Android
logging framework)