Create TimerDataInterface
In this PR we are creating TimerDataInterface, a common interface
between TimerData and ScopeTreeTimerData (based on ScopeTree). This
interface will be used by TimerDataProvider to easily communicate with
both classes.

This PR adjusted both classes to the common interface:

ScopeTreeTimerData:
- AddTimer now includes also an unused depth.

TimerData:
- Refactorized to have GetTimerMetadata().

- Created but not implemented GetTimers(): Should include some
optimization to discard several timers who after will occupt same
position in screen to not produce a regression when zooming-out.

- Added GetLeft/Right/Top/Down methods.

- Added unused methods: GetThreadId and OnCaptureComplete.

- Make some methods private.

- TODO: Create TimerData.cpp because TimerData is too big.

Tracks:

- Adapted to use GetTimerMetadata instead of GetMaxDepth() and
IsEmpty(), etc. Also refactorized a bit to use internal methods when
possible.

This PR is part of the process to move all timer-related data outside
the UI (http://b/202110356).

TODO:
- Implement GetTimers in TimerData (http://b/204173236).
- Use Provider in each TimerTrack (http://b/204173236).
- Improve TimerDataTests (http://b/204173036).