Create TimerDataInterface
In this PR we are creating TimerDataInterface, a common interface
between TimerData and ScopeTreeTimerData (based on ScopeTree). This
interface will be used by TimerDataProvider to easily communicate with
both classes.

In the future we might be interested in replace GetChains by proper
queries (to access specific timers of some function, and a few more
uses), but the main queries are already here.

This PR is part of the process to move all timer-related data outside
the UI (http://b/202110356).

Henning's already given LGTM in
(https://github.com/google/orbit/pull/2933)