Create conan recipe for libunwindstack's dependencies (#2941)
This is adding a new recipe which contains all the Android libraries
that our fork of libunwindstack needs. These notably are `libbase`,
`libprocinfo`, and `liblog`.

It also removes the no longer needed `libunwindstack` recipe.

This commit does not yet integrate our fork into the buildsystem. This
will follow in another commit.