Make libunwindstack GCC-compatible (#2939)
This commit applies the patches that we used to have in the
libunwindstack conan recipe to the fork.

`libunwindstack` only promises to compile with the Android standard
compiler which is clang and that results in a couple of compilation
errors when compiled with GCC.

Just to be clear: All the changes made here are making the code more C++
standard compliant. It's not that we have to introduce workarounds to
make GCC work.