Add UserSpaceFunction{Entry,Exit}PerfEvent to LinuxTracing
We are going to process function entries and exits (`FunctionEntry`,
`FunctionExit` protos) emitted by user space instrumentation in `LinuxTracing`,
in the same way we process uprobes. See the prototype of the integration of user
space instrumentation and unwinding as per
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm
These `struct`s represent those function entries and exits.

Notes:

These events are not actually produced yet, as the actual integration with user
space instrumentation will come as a later change.

I'm renaming variables, fields, comments to generalize from uprobes to any
dynamic instrumentation (uprobes or user space), but I'm not renaming the
classes that start with `Uprobes` just yet, mostly because we need to come up
with good names.

Detection of whether a sample fell inside a user space instrumentation
trampoline, or whether an address is inside a user space instrumentation return
trampoline for frame pointer callchain patching, is another matter (and quite
more complex) and will come as a later change.

Bug: http://b/194704608

Test: Unit tests. Tried as part of the integration of user space instrumentation
and unwinding.