Adapt TimerData to TimerDataInterfacei
In this PR we are adapting TimerData to TimerDataInterface, the common
interface between TimerData and ScopeTreeTimerData (based on ScopeTree).

Basically, we:

- Refactorized to have GetTimerMetadata().

- Created but not implemented GetTimers(): Should include some
optimization to discard several timers who after will occupt same
position in screen to not produce a regression when zooming-out.

- Added GetLeft/Right/Top/Down methods.

- Added unused methods: GetThreadId and OnCaptureComplete.

- Make some methods private.

And created 2 following bugs.

This PR is part of the process to move all timer-related data outside
the UI (http://b/202110356), and has been reviewed by Henning in
https://github.com/google/orbit/pull/2933.