Add seccomp profile for libunwindstack test runs (#2949)
libunwindstack unit test runs need the process_vm_readv syscall
which is disabled in Docker containers by default using the Linux
kernel's seccomp mechanism. But it's possible to provide a custom
seccomp profile.

So I took the Docker default profile from
(https://docs.docker.com/engine/security/seccomp/)
and added the `process_vm_readv` syscall to the unrestricted section.

I also added the option to the `build_and_upload_dependencies.sh` script
that also uses Docker. But I couldn't get it to work with the same
profile - only by setting the seccomp value to `unconfined`. I believe
this is caused by a bug that appears in the Docker version on our
workstations. I will revisit that as soon as it's fixed. Since the
script will only ever be executed manually this is not a big deal.