Replace `int ordered_in_file_descriptor` with `PerfEventOrderedStream` (#2945)
In addition to `PerfEvent`s sorted in their perf_event_open ring buffer, we want
to support events sorted in their thread. This is for the integration of user
space instrumentation and unwinding as per
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm

We add the `class PerfEventOrderedStream` that represents what stream
`PerfEvent`s are already ordered in (if any), and we update `PerfEvent` and
`PerfEventQueue` accordingly.

We compared performance with the current solution and couldn't note any
measurable difference.

Bug: http://b/194704608

Test: Expand unit tests. Run `LinuxTracingIntegrationTests`. Capture Trata with
all capture options enabled.