Add DIA SDK license to NOTICE file
This is adding the license agreement that applies to the DIA SDK to our
NOTICE file. It is the same as the one for Visual Studio which makes it
a bit misleading, especially since the DIA SDK is not mentioned anywhere
in the license text. But it has been checked and approved from other
sides, so we are just following this guidance here.