Allow the Vulkan layer to dump the current PID (#2932)
* Allow the Vulkan layer to dump the current PID

Write the PID to a file when the enviroment variable
ORBIT_VULKAN_LAYER_PID_FILE is set. This will allow
automated tools to have the required information to
initiate captures without being coupled to the
application.

The ORBIT_VULKAN_LAYER_PID_FILE will contain the file
where the PID should be written to.

Bug: http://b/204173805
Test: Game run with/without the envar, valid/invalid
envar value. Three unit tests have been created, one
for writing a PID to a valid file, check that it fails to
write a file to a nonexistent path, check that it fails
to write to an invalid filename, like a directory.