Adapt ScopeTreeTimerData to TimerDataInterface
In this PR we are adapting ScopeTreeTimerData to TimerDataInterface,
the common interface between TimerData and ScopeTreeTimerData.

For that we only need to add an unused_depth argument to AddTimer() and
erase the const in GetTimerMetadata().

This is the last part of https://github.com/google/orbit/pull/2933,
which was already reviewed by Henning.