Rename kUprobesPatchingFailed to kCallstackPatchingFailed
We will use this error also in the case of user space instrumentation so I want
to move away from a name that contains "uprobes".
I'm not creating a new error only for user space instrumentation because in the
general case of both uprobes and user space instrumentation it wouldn't be
trivial (or possible at all) to "blame" one or the other.

Bug: http://b/194704608

Test: Build.