Support optimized queries in ThreadTrackDataProvider
In this PR we are extending ThreadTrackDataProvider to use the new
method in ScopeTreeTimerData which allows the optimization
made in https://github.com/google/orbit/pull/2780.
The new method will be used by ThreadTracks to draw their timers.

This PR is a puzzle piece of making ThreadTracks to use the provider
instead of their own data (#2891)
which aims to solve http://b/202110929.

Ps: This PR will be rebased on
https://github.com/google/orbit/pull/2960 (and therefore TimersInTest::
will dissapear).