Use addresses and map name of user space instrumentation in LinuxTracing
This should be the last part that makes unwinding work with user space
instrumentation, in particular regarding error handling.

We are passing the address ranges of user space instrumentation trampolines to
`LinuxTracing` via the `UserSpaceInstrumentationAddresses` interface. We are
also passing the map name of the injected library. We now use these for the
following purposes:

- Detecting if a sample fell directly in a trampoline (entry or return), in
  `UprobesUnwindingVisitor`;
- Detecting if a sample fell in the injected library, in
  `UprobesUnwindingVisitor`;
- Detecting if unwinding stopped at a return trampoline, in
  `UprobesUnwindingVisitor`;
- Patching frame pointer callchains in `UprobesReturnAddressManager`.

Those classes are extended to now consider user space instrumentation in
addition to uprobes.
Variables and fields are renamed to generalize from uprobes to instrumented
functions, and comments are adjusted for the same reason.
Unit tests are extended to consider the new cases.

Note that with uprobes the approach was sometimes different and a bit simpler,
as most of the work for uprobes happens in the kernel, for which we don't get
samples, and for the rest we could simply use the map name "[uprobes]",
corresponding to a couple of addresses for the return trampoline.

Regarding performance: When user space instrumentation is enabled, we now
iterate over every frame of every callstack, and for each we get the
corresponding map entry. Considering a reasonable amount of samples per second,
and a reasonable amount of frames, this still sounds fine.
I looked at possible differences in CPU utilization on Trata while sampling at
5000 Hz (DWARF unwinding) and instrumenting a function with user space
instrumentation.
When instrumenting `item_coordinate_frame::update`, which is called often, this
change makes CPU utilization actually go down a bit, most likely because several
samples are discarded earlier than usual, causing less processing.
When instrumenting a function that is never called, I couldn't measure any
significant difference. I even profiled `OrbitService` with another instance of
Orbit for this case and indeed the number of samples falling in the new code is
extremely limited.

Bug: http://b/194704608
Bug: http://b/204404077

Test: Unit tests. Tried as part of the integration of user space instrumentation
and unwinding as per
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm,
on Trata, on a custom target, and on the same custom target also with frame
pointers.