Create abstract base class for object files such as ELF and COFF files
As a prerequisite for implementing unwinding for PE/COFF frames, we
create an abstract base class that is implemented by the `Elf` class
(and will be implemented by `Coff` in a future commit).

This is a purely mechanical change that declares all methods needed to
be implemented by a class that is returned from a `MapInfo` instance in
the `Object` base class.

Tested: No new code implemented, only interface changes. Nevertheless,
I built the service and tested on a sample application if all works.
Bug: http://b/192514457