Make ThreadTracks query data from DataProvider (#2969)
* Make ThreadTracks query data from DataProvider

This PR finally switch ThreadTracks to query data from
ThreadTrackDataProvider instead of owning it.

All the methos to query the data are already properly tested in previous
PRs (in ThreadTrackDataProviderTest and ScopeTreeTimerDataTests).

As a result we are solving http://b/202110929.

As a summary:
- We are replacing all references to TimerData and ScopeTree in
ThreadTracks per queries to the provider (overriden several methods from
TimerTrack).

- CaptureData doesn't need to have ThreadTrack chains anymore in
OnCaptureComplete().

A real analysis of performance is needed. Enforcing a full redraw is
every frame is showing me 4%/5% improvement in time to Draw +
UpdatePrimitives, but this is not very precise.

There are two big TODOs:
- Create ThreadTracks only before drawing and not when processing a
Timer (in capture thread). (http://b/198135618)
- Expand ThreadTrackDataProvider to TimerDataProvider which provides
timer queries to all TimerTracks. Work in progress, we are probably in
70%. (http://b/202110356)

* Fixup: Address virtual/override issue

* Fixup: ThreadTrack test construct in site