Generalize MapInfo to use Object instead of Elf and rename appropriately
In order to use different types of object files, we have already
introduced an more general `Object` class. This class is now also
used in `MapInfo`, which is generalized to general object file types.
This unfortunately triggers a large number of required renaming and
changes throughout the library.

* The main changes are in `MapInfo` and `Unwinder`, using `Object` and
renaming a couple of fields from "elf" to "object" to generalize.
* The `Elf` cache is moved up to the `Object` class. This is mainly
moving code from Elf.cpp to Object.cpp. This could probably be its own
change, but it felt inconsistent without this step.
* Remaining changes (in tests and Regs* classes) are just changes
triggered by renaming.

Tested: Unit tests; built version and ran on a sample application.
Bug: http://b/192514457