Fix UnwinderTest when running locally
UnwinderTest.build_frame_pc_in_jit failed when running locally, as
earlier tests overwrote the fake memory when all tests were run in the
same process (which is typically not the case in continuous
integrations). We fix this by making `memory_`, `process_memory_` and
(to be careful) `regs_` instance members of each test fixture and
initializing them in the constructor (so they are freshly created
and initialized for each TEST_F).

Tested: Ran unit tests locally.