Enable LiveFunctionsDataView to export function calls to CSV
With this change, we enable the live tab to support exporting all
function call events associated with the selected functions into a CSV
file.

The exported information contains the function name, thread name & id,
start timestamp, end timestamp, duration. And this action "Export Events
to CSV" is added to the same action group as "Copy Selection" and
"Export to CSV".