Pass addresses and map name of userspace instrumentation to LinuxTracing (#2984)
We are passing the address ranges of user space instrumentation trampolines to
`LinuxTracing`. We are also passing the map name of the injected library. We
need these for the following purposes:

- Detecting if a sample fell directly in a trampoline (entry or return), in
  `UprobesUnwindingVisitor`;
- Detecting if a sample fell in the injected library, in
  `UprobesUnwindingVisitor`;
- Detecting if unwinding stopped at a return trampoline, in
  `UprobesUnwindingVisitor`;
- Patching frame pointer callchains in `UprobesReturnAddressManager`.

We hold the address ranges and the injected library's map name with the
`UserSpaceInstrumentationAddresses` interface.

Actual usage of this information will come as the next change.

Note that with uprobes the approach was simpler, as most of the work for uprobes
happens in the kernel, for which we don't get samples, and for the rest we could
simply use the map name "[uprobes]", corresponding to a couple of addresses for
the return trampoline.

Bug: http://b/194704608
Bug: http://b/204404077

Test: Unit tests. Tried as part of the integration of user space instrumentation
and unwinding as per
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm,
also with frame pointers on custom target.