Add parsing for PE/COFF headers in libunwindstack
We add parsing of headers in PE/COFF files to our libunwindstack fork
in preparation for unwinding through PE/COFF frames. The parser is
skipping over some parts that we don't strictly need, such as several
members of the DOS header. Some other parts that we need are still
missing: We do not yet parse the section headers, which we will need to
extract the unwind information and other critical data. This wil be
added in a later change.

Tested: Unit tests.
Bug: http://b/192514457