Unify time graph (#2853)
This finally makes the TimeGraph implement the CaptureViewElement API correctly: It exposes its children through GetChildren and GetVisibleChildren, and uses the shared implementation of the super class to update and draw all tracks.

A lot is happening in this PR as I needed to remove a bunch of special treatment from the TimeGraph:

- Some parameters that were formerly passed during rendering calls have been converted to UI element properties, must notably the z offset and indentation. This follows the reasoning that all layout-affecting changes should happen before Draw
- z-offset is now using the batcher transformation stack
- The UpdatePrimitives method of CaptureViewElement is now protected as it requires knowledge that should not be needed from outside. Only Draw needs to be called from outside, and it is the responsibility of TimeGraph to determine if UpdatePrimitives is needed as it also owns the Batcher in this case
- CaptureViewElement::UpdatePrimitives is optimized to only draw children on screen
- There is no distinction between redraw_needed and update_primitives_required inside TimeGraph anymore as this wasn't used for some time. The main distinction now is whether the TimeGraph needs an update or not, and this triggers UpdatePrimitives if needed.

Finally, the main change returns all tracks as result of TimeGraph::GetChildren and TimeGraph::GetVisibleChildren, and gets rid of the special treatment that tracks have received as implicit children of TimeGraph.

Next steps: Slowly, slowy we'll be able to add unit tests for this. I'm planning on removing most of the special handling for accessibility classes (exposing the track tab as an actual visual element). I should also add a doc that describes the status quo...

Bug: b/176216022
Test: Multiple manual tests to confirm all still works correctly: Opening a larger capture, expanding / collapsing tracks, making sure the rendering looks like before. Tested track reordering, filtering, track type visibility, iterators, right-click measurement for visual correctness, and for rendering performance regressions.