Support format of sched:sched_wakeup before and after 5.14 kernel
The format of this tracepoint change with kernel version 5.15.
Don't assume a fixed layout, but only read the first fields: those are fixed and
we only need those.

Add dedicated `ConsumeSchedWakeupPerfEvent` method.
Also add a couple of `CHECK`s for other tracepoints.

Bug: http://b/205284110

Test:
- Run LinuxTracingIntegrationTests on 5.10 and 5.15 (was crashing on 5.15).
- Capture Trata on 5.10 and 5.15 with thread state collection enabled
(`OrbitService` was crashing on 5.15), check that thread state makes sense.