Fix the resizing behavior in process list table
With this change, we fixed the resizing behavior in the process list
table. Now, the column sizes in the process list cannot be changed by
the user or programmatically. The size of the PID column and the CPU
column are adjusted automatically according to the content; and the size
of the Name column is adjusted automatically to fill the available
space.

Bug: http://b/205556055