Fix the resizing behavior in the recent captures table
With this change, we fixed the resizing behavior in the recent captures
table. Now, the column sizes in the process list cannot be changed by
the user or programmatically. The size of the Last Used column and the
Created column are adjusted automatically according to the content; and the
size of the Filename column is adjusted automatically to fill the available
space.

Bug: http://b/205556055