Improve context menus for data views (#2986)
* Add DataView::GetContextMenuWithGrouping

With this change, we add the `DataView::GetContextMenuWithGrouping`
method which return a vector of action groups instead of a vector of
actions. This enables adding actions to view tabs by group.

Bug: http://b/193604470

* Add CallstackDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the callstack tab and implement
the GetContextMenuWithGrouping method for CallstackDataView.

Bug: http://b/193604470

* Add FunctionsDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the functions tab and implement
the GetContextMenuWithGrouping method for FunctionsDataView.

Bug: http://b/193604470

* Add LiveFunctionsDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the live tab and implement the
GetContextMenuWithGrouping method for LiveFunctionsDataView.

Bug: http://b/193604470

* Add ModulesDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the modules tab and implement
the GetContextMenuWithGrouping method for ModulesDataView.

Bug: http://b/193604470

* Add PresetsDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the presets tab and implement
the GetContextMenuWithGrouping method for PresetsDataView.

Bug: http://b/193604470

* Add SamplingReportDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the sampling tab and implement
the GetContextMenuWithGrouping method for SamplingDataView.

Bug: http://b/193604470

* Add TracepointsDataView::GetContextMenuWithGrouping

With this change, we reorder actions for the tracepoints tab and implement
the GetContextMenuWithGrouping method for TracepointsDataView.

Bug: http://b/193604470

* Update orbittreeview to use GetContextMenuWithGrouping

With this change, we changed orbittreeview to use
GetContextMenuWithGrouping instead of GetContextMenu. Also separators
are added among action groups.

Bug: http://b/193604470

* Update data view tests to test GetContextMenuWithGrouping

With this change, we updated tests for the data views after changing to
use GetContextMenuWithGrouping.

Bug: http://b/193604470

* Remove GetContextMenu for data views

With this change, we removed the GetContextMenu method for data views.

Bug: http://b/193604470

* [fixup] put "Add iterators(s)" and "Jump to" actions in a new group

* [fixup] address review comments

* [fixup] add CHECK for selected_indices