Add OrbitServiceIntegrationTests with a couple of initial tests
The idea is similar to `LinuxTracingIntegrationTests` but covering
`OrbitService` as a whole from the gRPC `CaptureService` level.
We start with a basic smoke test and a simple test for user space
instrumentation.
Some code from `LinuxTracingIntegrationTests` is factored out and reused.
I'm not renaming the module (directory name and namespace) for now.

Bug: http://b/206081462

Test: Run `OrbitServiceIntegrationTests` a hundred times. Also run
`LinuxTracingIntegrationTests`.