Create LockFreeUserSpaceInstrumentationEventProducer in StartNewCapture
While it's normal that data at the beginning or end of a capture cannot be
perfectly accurate, user space instrumentation missed the first function call to
an instrumented function even when this happens long after the library has been
injected and the capture has started.
This is because the connection to OrbitService was opened on the first entry
trampoline, at which point it is not yet ready to send data.
We solve this by establishing the connection on the start of the first capture.
Some care needs to be taken because this happens while all threads of the target
process are stopped, but gRPC spawns new threads, which now need to also be
stopped.

Bug: http://b/205939288

Test: User space instrumentation's unit tests. Run upcoming
`OrbitServiceIntegrationTests`, during whose setup this was discovered. Capture
Trata when instrumenting with user space instrumentation.