Reporting of status info from user space instrumentation.
Occationaly it is not possible to instrument a function with user space
instrumentation. We fall back to uprobes in that case.

Previously this was only reported into the log of OrbitService. Since
this has performance implications we'd like the user to see warnings in
this case. So this PR adds proper reporting into the capture log.

Bug: http://b/195543569
Test: Local runs, unit tests in the PR.