Use libunwindstack library function to convert errors to strings
In the recent libunwindstack version we are now using, additional
error codes have been added that cause issues for our own conversion
method to convert error codes to strings.

However there is now a library function in unwindstack/Error.h that we
can use to do the conversion. The converted strings are different from
what we had before, for example, "ERROR_NONE" is now "None".