Limit the maximum granularity of the time interval (#3027)
* Limit the maximum granularity of the time interval

Discussed once in the daily meeting. We agreed to have minimum 1 ns per
pixel to avoid visual issues but after seeing that we could have some
function calls of 50ns duration and some discussions, I decided to limit
100ns to the whole visible capture. If we try to zoom to a smaller
interval, we will center that interval in the minimum time range.

This PR is moving that new limit to the SetMinMax function to be
consistent. Before we had:

- 1 ns limit for zooming with mouse wheel
- no limit with ctrl + mouse right click

We have some issues when zooming-in a lot (Timers: http://b/168283058,
ThreadStates: http://b/200905941, and I even have a crash with ctrl + mouse
right click). They won't appear anymore after this PR.

PS: As a limitation, the overlay who is generated when holding mouse right
click is still rounded to an integer nanoseconds and this rounded is still
visible with 100ns total time range, as we can see in the screenshot:
http://screen/3x9qPTZWJsjnEQo. Anyway, it's not too drastic with 100ns and
I will consider that expected behaviour.

Test: several interactions in a loaded capture.

* fixup: const