Create LockFreeApiEventProducer in orbit_api_set_enabled (#3035)
Orbit Api missed at least the first event even if this came a long time after
`liborbit.so` was injected and `orbit_api_set_enabled` was called for the first
time. This is because the connection to `OrbitService` was opened on the first
event, at which point it was not yet ready to send data.
We solve this by establishing the connection in `orbit_api_set_enabled`, which
is called at the very beginning of a capture.
As this call happens while the target is stopped, but that call will now cause
gRPC to spawn new threads, these now also need to be stopped.

This is very similar to 72a078994335b8af3be7bfcc687045a9553f2ed5 fixing
http://b/205939288.

Bug: http://b/206359125

Test: Run the upcoming Orbit Api test from the upcoming
`OrbitServiceIntegrationTests`, during whose implementation this issue was
discovered. Capture OrbitTest with Orbit Api enabled.