Fix potentially wrong program counter reported by ORBIT_SCOPE
`__builtin_return_address` was previously used in `Scope::Scope`, but this
yielded different results depending on whether this constructor was getting
inlined or not (we were assuming it was never getting inlined).
So instead use inline assembly to get the instruction pointer before building a
`Scope`.

A simpler alternative would be to simply make `Scope::Scope` noinline, but we
don't want to prevent this optimization (that would have been much easier,
though...).

Note that `Orbit.h` is also included on Windows, even if OrbitApi is not
implemented there. For the Windows case, we simply prevent inlining as just
mentioned, as I wasn't able to find a different solution (for example, inline
assembly is 32 bit only).

Bug: http://b/206418525

Test: Run upcoming "OrbitApi" test case of `OrbitServiceIntegrationTests`, that
found out about this. Capture OrbitTest with manual instrumentation and verify
that tooltips of scopes are correct. Build on Windows (in particular OrbitTest,
which includes Orbit.h).