Parse section headers of PE/COFF object files
We need to parse the section headers to find data about some specific
sections we need for unwinding, in particular the .text and a
potentially present .debug_frame section. With this change we are
adding parsing of these headers.

We also add looking up section names in the string table, which is
required for section names that are longer than 8 characters.

Tested: Unit tests.
Bug: http://b/192514457